+++ src/mruby_integration/image.cpp
@@ -146,6 +146,15 @@ mrb_value mrb_image_crop(mrb_state *mrb, mrb_value) {
   return mrb_nil_value();
 }
 
+mrb_value mrb_image_alpha_mask(mrb_state *mrb, mrb_value) {
+  Image *image, *alpha_mask;
+  mrb_get_args(mrb, "dd", &image, &Image_type, &alpha_mask, &Image_type);
+
+  ImageAlphaMask(image, *alpha_mask);
+
+  return mrb_nil_value();
+}
+
 mrb_value mrb_image_resize_nearest_neighbour(mrb_state *mrb, mrb_value) {
   mrb_value image_obj;
   mrb_int width, height;
@@ -189,6 +198,7 @@ void append_images(mrb_state *mrb) {
   mrb_define_method(mrb, mrb->kernel_module, "image_from_image", mrb_image_from_image, MRB_ARGS_REQ(2));
   mrb_define_method(mrb, mrb->kernel_module, "image_text_ex", mrb_image_text_ex, MRB_ARGS_REQ(5));
   mrb_define_method(mrb, mrb->kernel_module, "image_crop!", mrb_image_crop, MRB_ARGS_REQ(2));
+  mrb_define_method(mrb, mrb->kernel_module, "image_alpha_mask!", mrb_image_alpha_mask, MRB_ARGS_REQ(2));
 
   mrb_define_method(mrb, mrb->kernel_module, "image_resize!", mrb_image_resize, MRB_ARGS_REQ(3));
   mrb_define_method(mrb, mrb->kernel_module, "image_resize_nearest_neighbour!", mrb_image_resize_nearest_neighbour, MRB_ARGS_REQ(3));